(This is a merge from branch CX_2_4_SUP - See Commit #22521])
Declare Language before invoking logger

Fix SR 219 on Partners
https://partners.xrce.xerox.com/tracker/?func=detail&aid=219&atid=199&group_id=120

+++ SF/www/account/logout.php
@@ -8,8 +8,9 @@
 
 require_once('pre.php');    
 
-db_query("DELETE FROM session WHERE session_hash='$session_hash'");
-
+if (isset($session_hash)) {
+    db_query("DELETE FROM session WHERE session_hash='$session_hash'");
+}
 session_cookie('session_hash','');
 session_redirect('/');
